requirements.txt
click==8.1.3
Flask==2.1.2
Flask-WTF==1.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.3
Pillow==9.1.0
Werkzeug==2.1.2
WTForms==3.0.1
